Create `latent_representation` class
I created an initial stab at a `latent_representation` class to store mu+sigma values for distributions over the latent space of a VAE. The class avoids using any torch/tensor stuff, so it should be totally compatible with the rest of the Soar codebase. However, this does mean converting to/from tensors must happen in the `torch_module_wrapper` class.

Also contains a few fixes for neural network classes.