Refactor VOp code to split parent params in `opencv_image` and `latent_representation` types.
In order to use the `latent_representation` type in VOps, I needed to allow VOp specs to specify that a parent node was providing an input latent or that the VOp provides an output latent. This required a bit of refactoring and some additions to various parts of the VOp pipeline.