Add `get-affine-from-corners` VOp, others, bugfixes
Bit of a large commit, but oh well. Adds code for the `get-affine-from-corners` VOp, which ended up not working as well as hoped. Also adds a couple other new VOps: `resize-image` and `scale-image`, and adds slight improvements to the `get-transforms` VOp via the object rep's `get_best_affine_transforms` method.